change vec loader

+++ progressivis/core/common.py
@@ -3,4 +3,4 @@ import numpy as np
 class ProgressiveError(Exception):
     pass
 
-NIL = np.array([])
+NIL = np.array([],dtype=int)
